printf

+++ _octal.c
@@ -7,7 +7,7 @@
 **/
 int _octal(va_list args)
 {
-	unsigned int n = va_arg(args, int);
+	unsigned long int n = va_arg(args, int);
 	char arr[200];
 	int i = 0;
 	int j;
